finished off other migrations
